requirements.txt
django
gunicorn
python-dotenv
django-crispy-forms
django_heroku
django_filter
dj-database-url
psycopg2
rollbar
